+++ server/src/api/socket/asset_manager/core.py
@@ -15,6 +15,7 @@ from aiohttp import web
 
 from .... import auth
 from ....app import app, sio
+from ....config import config
 from ....db.models.asset import Asset
 from ....db.models.asset_rect import AssetRect
 from ....db.models.user import User
@@ -346,6 +347,17 @@ async def handle_regular_file(upload_data: ApiAssetUpload, data: bytes, sid: str
     return asset_dict
 
 
+@sio.on("Asset.Upload.Limit", namespace=ASSET_NS)
+@auth.login_required(app, sio, "asset")
+async def assetmgmt_upload_limit(sid: str):
+    user = asset_state.get_user(sid)
+    return {
+        "single": config.getint("Webserver", "max_single_asset_size_in_bytes"),
+        "used": user.get_total_asset_size(),
+        "total": config.getint("Webserver", "max_total_asset_size_in_bytes"),
+    }
+
+
 @sio.on("Asset.Upload", namespace=ASSET_NS)
 @auth.login_required(app, sio, "asset")
 async def assetmgmt_upload(sid: str, raw_data: Any):
@@ -363,6 +375,7 @@ async def assetmgmt_upload(sid: str, raw_data: Any):
     if uuid not in asset_state.pending_file_upload_cache:
         asset_state.pending_file_upload_cache[uuid] = {}
     asset_state.pending_file_upload_cache[uuid][upload_data.slice] = upload_data
+    # todo: verify that `totalSlices` does not change between messages
     if len(asset_state.pending_file_upload_cache[uuid]) != upload_data.totalSlices:
         # wait for the rest of the slices
         return
@@ -373,6 +386,21 @@ async def assetmgmt_upload(sid: str, raw_data: Any):
         data += asset_state.pending_file_upload_cache[uuid][slice_].data
 
     del asset_state.pending_file_upload_cache[upload_data.uuid]
+
+    total_asset_size = user.get_total_asset_size()
+    max_single_asset_size = config.getint("Webserver", "max_single_asset_size_in_bytes")
+    max_total_asset_size = config.getint("Webserver", "max_total_asset_size_in_bytes")
+
+    if max_single_asset_size > 0 and len(data) > max_single_asset_size:
+        logger.warn(
+            f"{user.name} attempted to upload a file that is too large ({len(data)} > {max_single_asset_size})"
+        )
+        return
+    if max_total_asset_size > 0 and total_asset_size + len(data) > max_total_asset_size:
+        logger.warn(
+            f"{user.name} attempted to upload a file that is too large ({total_asset_size + len(data)} > {max_total_asset_size})"
+        )
+        return
 
     return_data = None
     file_name = upload_data.name